+++ packages/default-client/src/index.ts
@@ -1,4 +1,4 @@
-import {ScheduleInfo, ScheduledExecution, TimeBasedTiming, RoundBasedTiming} from 'fuzd-scheduler';
+import {ScheduleInfo, ScheduledExecution, TimeBasedTiming, RoundBasedTiming, DecryptedPayload} from 'fuzd-scheduler';
 import {timelockEncrypt, HttpChainClient, roundAt} from 'tlock-js';
 import {privateKeyToAccount} from 'viem/accounts';
 import {BroadcastSchedule, TransactionSubmission} from 'fuzd-executor';
@@ -10,18 +10,6 @@ export type ClientConfig = {
 	schedulerEndPoint: string | ((execution: string, signature: `0x${string}`) => Promise<ScheduleInfo>);
 	privateKey: `0x${string}`;
 };
-
-// TODO share with decrypter
-export type DecryptedPayload<TransactionDataType> =
-	| {
-			type: 'time-locked';
-			payload: string;
-			timing: RoundBasedTiming;
-	  }
-	| {
-			type: 'clear';
-			transaction: TransactionDataType;
-	  };
 
 export function createClient(config: ClientConfig) {
 	const wallet = privateKeyToAccount(config.privateKey);
@@ -57,18 +45,20 @@ export function createClient(config: ClientConfig) {
 
 		const payloadJSON: DecryptedPayload<TransactionSubmission> = {
 			type: 'clear',
-			transaction: {
-				type: '0x2',
-				chainId,
-				gas: ('0x' + execution.gas.toString(16)) as `0x${string}`,
-				broadcastSchedule: execution.broadcastSchedule.map((v) => ({
-					duration: ('0x' + v.duration.toString(16)) as `0x${string}`,
-					maxFeePerGas: ('0x' + v.maxFeePerGas.toString(16)) as `0x${string}`,
-					maxPriorityFeePerGas: ('0x' + v.maxPriorityFeePerGas.toString(16)) as `0x${string}`,
-				})) as BroadcastSchedule,
-				data: execution.data,
-				to: execution.to,
-			},
+			transactions: [
+				{
+					type: '0x2',
+					chainId,
+					gas: ('0x' + execution.gas.toString(16)) as `0x${string}`,
+					broadcastSchedule: execution.broadcastSchedule.map((v) => ({
+						duration: ('0x' + v.duration.toString(16)) as `0x${string}`,
+						maxFeePerGas: ('0x' + v.maxFeePerGas.toString(16)) as `0x${string}`,
+						maxPriorityFeePerGas: ('0x' + v.maxPriorityFeePerGas.toString(16)) as `0x${string}`,
+					})) as BroadcastSchedule,
+					data: execution.data,
+					to: execution.to,
+				},
+			],
 		};
 		const payloadAsJSONString = JSON.stringify(payloadJSON);