+++ packages/gateways/CHANGELOG.md
@@ -1,5 +1,12 @@
 # fuzd-gateways
 
+## 0.1.13
+
+### Patch Changes
+
+- Updated dependencies
+  - fuzd-scheduler@0.2.1
+
 ## 0.1.12
 
 ### Patch Changes