+++ packages/scheduler/CHANGELOG.md
@@ -1,5 +1,11 @@
 # fuzd-scheduler
 
+## 0.2.1
+
+### Patch Changes
+
+- export DecryptedPayload + fix
+
 ## 0.2.0
 
 ### Minor Changes