+++ packages/tlock-decrypter/CHANGELOG.md
@@ -1,5 +1,13 @@
 # fuzd-tlock-decrypter
 
+## 0.2.1
+
+### Patch Changes
+
+- export DecryptedPayload + fix
+- Updated dependencies
+  - fuzd-scheduler@0.2.1
+
 ## 0.2.0
 
 ### Minor Changes